structure refactored
