+++ src/main/java/by/guesspic/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package spring_server.config;
+package by.guesspic.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
