+++ src/main/java/by/guesspic/controllers/RoomRequestController.java
@@ -1,31 +1,36 @@
-package spring_server.controllers;
+package by.guesspic.controllers;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import by.guesspic.data.Message;
+import by.guesspic.data.MessageType;
+import by.guesspic.data.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
-import spring_server.chat_model.*;
+import by.guesspic.room.Room;
+import by.guesspic.service.RoomActivityService;
+import by.guesspic.service.RoomService;
 
 @RestController
 public class RoomRequestController {
     @Autowired
-    private RoomActivityController activityController;
+    private RoomActivityService activityController;
     @Autowired
     private SimpMessageSendingOperations messagingTemplate;
 
     @Autowired
-    RoomsController roomsController;
+    RoomService roomsController;
 
     private static final Gson gson = new GsonBuilder().create();
 
     @PostMapping("new_user/{name}")
     public User addUser(@PathVariable("name") String name) {
         User user = new User(name);
-        int roomId = roomsController.addToRandomRoom(user);
+        int roomId = roomsController.addToNearestFreeRoom(user);
         activityController.updateActivity(roomsController.getRoomById(roomId),
                 user.getId());
 
@@ -49,7 +54,7 @@ public class RoomRequestController {
         return roomsController.getRoomById(roomId).getPainterId();
     }
 
-    @GetMapping("room/{roomId}/user/{painterId}/word")
+    @GetMapping("by/guesspic/room/{roomId}/user/{painterId}/word")
     public String getWord(@PathVariable("roomId") int roomId, @PathVariable("painterId") int painterId) {
         Room room = roomsController.getRoomById(roomId);
         if (room.getPainterId() == painterId) {
