+++ src/main/java/by/guesspic/dao/Dao.java
@@ -1,4 +1,4 @@
-package database;
+package by.guesspic.dao;
 
 import java.util.List;
 import java.util.Optional;
