+++ src/test/java/HibernateRequestTest.java
@@ -1,7 +1,7 @@
-import database.Dao;
-import database.HibernateRequest;
-import database.Word;
-import database.WordDao;
+import by.guesspic.dao.Dao;
+import by.guesspic.dao.HibernateRequest;
+import by.guesspic.data.Word;
+import by.guesspic.dao.WordDao;
 import org.junit.jupiter.api.Test;
 
 import java.sql.SQLException;
